+++ js/api.js
@@ -6,8 +6,8 @@ const Route = {
 };
 
 const ErrorText = {
-  GET_DATA: 'Не удалось загрузить данные. Попробуйте обновить страницу',
-  SEND_DATA: 'Не удалось отправить форму. Попробуйте ещё раз',
+  GET_DATA: 'Не удалось загрузить данные. Попробуйте обновить страницу.',
+  SEND_DATA: 'Не удалось отправить форму. Попробуйте ещё раз.',
 };
 
 const Method = {
