+++ js/draw-big-picture.js
@@ -3,11 +3,11 @@ import {isEscapeKey} from './util.js';
 const COMMENTS_NUMBER_STEP = 5;
 const bigPictureElement = document.querySelector('.big-picture');
 const bigPictureCloseElement = bigPictureElement.querySelector('.big-picture__cancel');
-const bigPictureImage = bigPictureElement.querySelector('.big-picture__img').querySelector('img');
-const bigPictureLikesCount = bigPictureElement.querySelector('.likes-count');
-const bigPictureDescription = bigPictureElement.querySelector('.social__caption');
-const bigPictureCommentsList = bigPictureElement.querySelector('.social__comments');
-const bigPictureCommentTemplate = bigPictureCommentsList.querySelector('.social__comment');
+const bigPictureImage = bigPictureElement.querySelector('.big-picture__img img');
+const bigPictureLikesCountElement = bigPictureElement.querySelector('.likes-count');
+const bigPictureDescriptionElement = bigPictureElement.querySelector('.social__caption');
+const bigPictureCommentsListElement = bigPictureElement.querySelector('.social__comments');
+const bigPictureCommentTemplate = bigPictureCommentsListElement.querySelector('.social__comment');
 const bigPictureCommentsLoaderElement = bigPictureElement.querySelector('.comments-loader');
 const bigPictureCommentsCountElement = bigPictureElement.querySelector('.social__comment-count');
 
@@ -15,23 +15,23 @@ let currentPicture;
 let currentCommentsIndex = 0;
 
 const clearCommentsList = () => {
-  bigPictureCommentsList.innerHTML = '';
+  bigPictureCommentsListElement.innerHTML = '';
 };
 
 const getCommentElement = (comment) => {
-  const formatedComment = bigPictureCommentTemplate.cloneNode(true);
-  const commentAutorImage = formatedComment.querySelector('.social__picture');
+  const commentElement = bigPictureCommentTemplate.cloneNode(true);
+  const commentAutorImage = commentElement.querySelector('.social__picture');
   commentAutorImage.src = comment.avatar;
   commentAutorImage.alt = comment.name;
-  formatedComment.querySelector('.social__text').textContent = comment.message;
-  return formatedComment;
+  commentElement.querySelector('.social__text').textContent = comment.message;
+  return commentElement;
 };
 
 const drawComments = () => {
   let commentIndex = 0;
   for (let i = currentCommentsIndex; i < currentCommentsIndex + COMMENTS_NUMBER_STEP; i++) {
     commentIndex = i;
-    bigPictureCommentsList.appendChild(getCommentElement(currentPicture.comments[i]));
+    bigPictureCommentsListElement.appendChild(getCommentElement(currentPicture.comments[i]));
     if (i + 1 === currentPicture.comments.length) {
       bigPictureCommentsLoaderElement.classList.add('hidden');
       break;
@@ -43,8 +43,8 @@ const drawComments = () => {
 
 const changeBigPicture = () => {
   bigPictureImage.src = currentPicture.url;
-  bigPictureLikesCount.textContent = currentPicture.likes;
-  bigPictureDescription.textContent = currentPicture.description;
+  bigPictureLikesCountElement.textContent = currentPicture.likes;
+  bigPictureDescriptionElement.textContent = currentPicture.description;
 
   clearCommentsList();
   drawComments();
@@ -70,7 +70,7 @@ const openBigPicture = (picture) => {
   bigPictureElement.classList.remove('hidden');
 };
 
-function closeBigPicture() {
+function closeBigPicture () {
   bigPictureElement.classList.add('hidden');
   document.body.classList.remove('modal-open');
   bigPictureCommentsLoaderElement.classList.remove('hidden');
